+++ src/components/Invoices/InvoiceDetail/InvoiceDetail.tsx
@@ -3,7 +3,8 @@ import { BaseDetailView } from '../../BaseDetailView/BaseDetailView'
 import { InvoiceForm } from '../InvoiceForm/InvoiceForm'
 import { UpsertInvoiceMode } from '../../../features/invoices/api/useUpsertInvoice'
 import { Heading } from '../../ui/Typography/Heading'
-import { InvoiceStatus, type Invoice, type InvoicePayment } from '../../../features/invoices/invoiceSchemas'
+import { InvoiceStatus, type Invoice } from '../../../features/invoices/invoiceSchemas'
+import { type InvoicePayment } from '../../../features/invoices/invoicePaymentSchemas'
 import { HStack, VStack } from '../../ui/Stack/Stack'
 import { DataPoint } from '../../DataPoint/DataPoint'
 import { Span } from '../../ui/Typography/Text'
