+++ src/components/Invoices/InvoicePaymentForm/useInvoicePaymentForm.ts
@@ -1,7 +1,8 @@
 import { useCallback, useMemo, useState, useRef } from 'react'
 import { revalidateLogic } from '@tanstack/react-form'
 import { useAppForm } from '../../../features/forms/hooks/useForm'
-import { type InvoicePayment, type DedicatedInvoicePaymentForm, UpsertDedicatedInvoicePaymentSchema, type Invoice } from '../../../features/invoices/invoiceSchemas'
+import { type Invoice } from '../../../features/invoices/invoiceSchemas'
+import { type InvoicePayment, type DedicatedInvoicePaymentForm, UpsertDedicatedInvoicePaymentSchema } from '../../../features/invoices/invoicePaymentSchemas'
 import { useUpsertDedicatedInvoicePayment, UpsertDedicatedInvoicePaymentMode } from '../../../features/invoices/api/useUpsertDedicatedInvoicePayment'
 import { Schema } from 'effect'
 import { convertInvoicePaymentFormToParams, getInvoicePaymentFormDefaultValues, validateInvoicePaymentForm } from './formUtils'
