+++ src/features/invoices/api/useUpsertDedicatedInvoicePayment.tsx
@@ -4,7 +4,8 @@ import useSWRMutation, { type SWRMutationResponse } from 'swr/mutation'
 import { post, put } from '../../../api/layer/authenticated_http'
 import { useLayerContext } from '../../../contexts/LayerContext'
 import { useAuth } from '../../../hooks/useAuth'
-import { InvoicePaymentSchema, InvoiceStatus, type Invoice, type InvoicePayment, type UpsertDedicatedInvoicePaymentSchema } from '../invoiceSchemas'
+import { InvoiceStatus, type Invoice } from '../invoiceSchemas'
+import { InvoicePaymentSchema, type InvoicePayment, type UpsertDedicatedInvoicePaymentSchema } from '../invoicePaymentSchemas'
 import { Schema, Effect } from 'effect'
 import { useInvoicesGlobalCacheActions } from './useListInvoices'
 import { useInvoiceSummaryStatsCacheActions } from './useInvoiceSummaryStats'
