+++ src/features/invoices/invoiceSchemas.ts
@@ -2,7 +2,6 @@ import { Schema, pipe } from 'effect'
 import { CustomerSchema } from '../../schemas/customer'
 import { InvoiceTermsValues } from '../../components/Invoices/InvoiceTermsComboBox/InvoiceTermsComboBox'
 import { ZonedDateTimeFromSelf } from '../../utils/schema/utils'
-import { PaymentMethodSchema, TransformedPaymentMethodSchema } from '../../components/PaymentMethod/schemas'
 
 export enum InvoiceStatus {
   Voided = 'VOIDED',
@@ -300,51 +299,3 @@ export const InvoiceSummaryStatsResponseSchema = Schema.Struct({
   ),
 })
 export type InvoiceSummaryStatsResponse = typeof InvoiceSummaryStatsResponseSchema.Type
-
-export const UpsertDedicatedInvoicePaymentSchema = Schema.Struct({
-  amount: Schema.Number,
-
-  method: PaymentMethodSchema,
-
-  paidAt: pipe(
-    Schema.propertySignature(Schema.Date),
-    Schema.fromKey('paid_at'),
-  ),
-
-  referenceNumber: Schema.optional(Schema.String).pipe(
-    Schema.fromKey('reference_number'),
-  ),
-
-  memo: Schema.optional(Schema.String),
-})
-
-export type UpsertDedicatedInvoicePayment = typeof UpsertDedicatedInvoicePaymentSchema.Type
-
-export const DedicatedInvoicePaymentFormSchema = Schema.Struct({
-  amount: Schema.BigDecimal,
-
-  method: Schema.NullOr(PaymentMethodSchema),
-
-  paidAt: Schema.NullOr(ZonedDateTimeFromSelf),
-
-  referenceNumber: Schema.String,
-
-  memo: Schema.String,
-})
-export type DedicatedInvoicePaymentForm = typeof DedicatedInvoicePaymentFormSchema.Type
-
-export const InvoicePaymentSchema = Schema.Struct({
-  amount: Schema.Number,
-
-  method: TransformedPaymentMethodSchema,
-
-  at: Schema.propertySignature(Schema.Date),
-
-  referenceNumber: pipe(
-    Schema.propertySignature(Schema.NullOr(Schema.String)),
-    Schema.fromKey('reference_number'),
-  ),
-
-  memo: Schema.NullOr(Schema.String),
-})
-export type InvoicePayment = typeof InvoicePaymentSchema.Type
